Güncelle
sınıfı
